Restructure the way we start and configure metro (#19)
This PR changes the way we interact with metro:

Before, we'd have a JS script that'd be a copy of what CLI code does
when starting metro. This resulted in compatibility issues as this
script changes between versions.

The new approach is to use react-native scripts directory to start
metro, and in order to inject our things, we only use the metro config
API, which is relatively stable. We provide the starter script with path
to our config, which in turn loads the original metro config for the
app, and then overrides settings that we wanted.