Better restarts handling (#39)
This PR improves the way we restart things.

1. When restart button is clicked in the UrlBar or on the error dialog,
we try to be smart about restarting. When the app is running we only
perform metro reload which is the quickest. If there was an error, we
try to run the project again
2. For complex issues, we introduced a new button for triggering clean
builds which starts a new native build and also clears metro cache.

Since changes to UrlBar component where necessary for it to talk using
new project api object, I also refactored navigation related commands to
use the api object approach.