Use qt-hide-window instead of no-window for Android emulator (#58)
no-window option seem to mess stuff up with emulator process. When set,
emulator won't ever get to 60fps and on some computers it'd be very slow
with high cpu usage and ~20fps. I don't understand the reason but
qy-hide-window seem option works ok in our case (it results in emulator
not displaying its native window) but doesn't have the downsides of the
no-window option