Don't run all builds by default and make them cancellable (#65)
This PR changes the way we were handling builds before. Before we'd
launch builds on startup and keep single "instance" of a build around in
the project class. This change is to prevent issues with early users
where initial builds were consuming too much resources.

Note that due to this change, if you end up switching between devices of
the same type, we will trigger a new build. We may want to optimize this
in the future.

This PR also introduces a mechanism for cancelling builds. This can
happen when you switch between devices while the build is ongoing. We
dont support full cancellations and only utilize subprocess.kill here in
order to kill spawned build processes that are taking long time to
complete. Other parts of the build process may continue to work even
after cancel is triggered.

As a part of the PR I'm also making some related/unrelated adjustments:
- I made it so that we refetch images list when the image dialog or the
device creation dialog is opened
- I fixed clean build flag such that it also triggers clean native build