Fix extension activation for some category of RN project configurations (#66)
This PR addresses a problem with extension not activating in certain
project setups. What we are changing here is:
1) we add new activation events, such that the extension can activate
for app.json and metro.config.js.
2) we allow for app root to not be the direct workspace root. On
extension acivation we search for the location of app root
3) we add additional diagnostics command that's available even when
extension is not active. This should allow people with unsupported
project schemes to learn why the extension can't be used.