Use subprocess.catch instead of on exit event to include errors when process cannot be started at all (#67)
This PR changes way we communicate back subprocess errors for some
long-running tasks.

I noticed that when there is a startup error (i.e. requested script or
binary doesn't exists), we don't get "on exit" event. In that scenario
we wouldn't surface task execution error and the promise that we return
would be unresolved forever. With this change we are updating all uses
of "on exit" with subprocess.catch that is triggered when process exists
with non-zero code, or when there is another execution error that
prevents the subprocess from launching