Update error handling around subprocesses to avoid excessive logger output (#84)
This PR updates the code related to handling errors for subprocess/exec
calls.

It addresses two issues:
1) Problem with commands we expect to return non-zero exit code
producing error output in logs
2) Problem with some of the commands being terminated as expected and
still producing unhandled promise rejection errors

Both of those errors would pollute log output making it more difficult
to track real problems so we are fixing it here.

To address issue (1) we add flag to exec calls that indicates that we
expect command to exit with non-zero code. Based on that flag exec
utility will avoid printing errors when they happen.

The problem (2) was due to a bug where we'd call `then` on subprocess
promise that may throw and after `then` call there was no `catch`
suggesting we may miss the rejection. However, in all those instances in
the codebase we already have a `catch` section, but it wasn't chained
with `then` resulting in the runtime thinking the error wasn't handled.
We fix this by chaining `catch` and `then`