Fix errors printed by devtool operation scanning code (#86)
This PR removes errors produced in the app by the code for scanning
devtools operations.

This code is used for the jump-to-route functionality, but was producing
errors that were shown in the user's app console, which is undesirable.

As part of this PR we are also changing the extraction algorithm in that
code, such that it only filters ADD operations and only ones that impact
non-root elements. There is no point in analyzing or inspecting other
operations as only additions would give us a good signal as to what new
elements appear when a screen gets mounted. We also can't analyze root
element operation as inspecting it always throws an error