Kinda one-way progress bar (#90)
The PR makes the problem of "going back" progress bar less severe.

Fully fixing the problem of progress bar going back is more complicated
as it would require tracking previous values of progress and turning off
the animation accordingly. The biggest problem I've encountered is the
progress bar unmounts and mounts a couple of times when building an app
make is hard to track the progress values. Definitely doable to fix this
properly but I don't think worth the effort right now.

### Before


https://github.com/software-mansion-labs/react-native-sztudio/assets/39658211/70d8d665-968a-40b3-844d-3da4565114ee



### After


https://github.com/software-mansion-labs/react-native-sztudio/assets/39658211/af864330-ba16-43f7-8061-c4fdacd2dc55