Patch Android builds with fixes allowing debugger to connect using port settings from shared preferences (#94)
This PR implements a workaround for Android builds in order to patch
them with a fix that's now merged upstream
https://github.com/facebook/react-native/pull/42617

The issue here is that we rely on the mechanism of React Native on
Android which reads settings from shared preferences in order to locate
metro port. However, React Native Android's implementation uses separate
code path for locating the port for the debugger, which normally is the
same as the metro port as debugger proxy is one of metro's endpoints.
That code path has been broken for a long time as it does not respect
the settings you can have in shared preferences, instead it only relies
on the port that's provided for the build. This isn't ideal, as we'd
need to make a separate build for each device session in the IDE.

While this is fixed upstream, we still want to support users on current
versions of React Native and Expo. Because in typical workflows, on
Android, the java code is not build from source but instead is loaded
from pre-built AAR files, we need to override the classes from that
package. The way we do it relies on injecting a so-called init script
with gradle CLI. The init script sets up an artifact transform which
hijacks the JAR file with classes unpacked from react-android
dependency, and repackages them while deleting some classes we want to
override. On top of that, we add a separate source directory that
includes sources for the missing classes. This way we are able to
override any internal class from React Android package.

We utilize this method to swap implementation of
`PackagerConnectionSettings.java` where the only change we're making is
to call `getDebugServerHost()` directly from `getInspectorServerHost()`
method. This way, the host for debugger uses the same location/port as
metro server.