Fix issue with not loading source file during uncaught exception (Android) (#139)
This PR fixes problem with not loading source file during uncaught
exception. The problem is resolved by modifying the hostname and port
used on Android emulator network. We change it to the same values used
by metro, so during script urls comparison, there is no mismatch between
hosts and ports.

Resolves #138