Update node modules installation check to npm exclusively (#14)
This PR restricts listing installed node modules to `npm` only. At the
same time it fixes the issue which marke node modules as installed even
if they were not present.

Before:
The check to verify packages was performed using different package
managers.

After:
Now we use only `npm` to determine the presence of node modules. `npm`
is included with `node` so there's no risk that user won't have it.