Updates to support expo-router v3 (#77)
This PR adds support for expo-router v3.

We add new plugin for v3 and rename the previous one to
expo_router_v2_plugin

The v3 plugin is almost identical to v2 as the APIs we rely on didn't
change. The only difference is in the location from which we require
router-store. In v2 it was under src while in v3 we require it from
build folder. Apparently, there is no good way to do a runtime check as
if the file doesn't exists in one version metro would report an error.

We decided to duplicate the code here as will likely phase out v2
support in near future and at that point we can remove the plugin
entirely.