Fix ios scheme detection (#99)
#90 introduced a regression due to improper use of basename function.
Because no extension was provided, basename would contain .xcworkspace
extension breaking iOS builds. On top of that, there was another bug
related to xcscheme search making it so that instead of including scheme
files we were excluding them.

This PR fixes the two above issues. We now pass an extension when using
basename and also use RelativePattern when searching for xcscheme files.