Restore expo env plugin (#100)
This PR re-enables expo env plugin allowing it to function.

The reason why we disabled it was that metro wasn't handling source maps
correctly. This was because env plugin would add additional modules to
output using customSerializer, but the serializer isn't used when
generating sourcemaps, and hence the source maps were generated by
skipping that module. As a consequence, all lines transformed using
sourcemaps were shifted by the number of lines of the env module.

In this PR we implement a different workaround that stores the number of
lines from expo env in global variable and then uses it to manually
adjust reported line numbers.