Fix inspector highlight on mouse move (#117)
In #113 we introduced a regression that broke inspector highlight when
hovering.

The problem was due to the fact that we'd call `stack.filter` in
`wrapper.ts` and when hovering, we never ask for the stack to be
populated. As a result we've been calling filter on undefined stack
which resulted in the callback never returning with the frame data.

This PR fixes the issue and moves the code that filters non-empty stack
items to the code branch that runs only when stack is requested.