Fix communication between webview and extension code. (#114)
This PR fixes a longstanding issue with leaking memory due to RPC calls
with callbacks from webview to main extension process.

There were two separate issue:
1) We'd create a new function even when attempting to send the same
callback method. As a consequence, all removeListener calls would never
work as we'd try to remove a new closure instead of using the previously
registered one.
2) We'd never cleanup a map of callbacks, because we couldn't tell if a
given callback is still being referenced on the main extension process
side.

This PR fixes these two problems:
1) We now keep a registry with WekRefs pointing to callback closures.
When callback with a known id is requested, we return one from that map.
This way, calls with same functions made on the webview side, will also
receive same objects on the extension side therefore allowing for
listeners to be deleted.
2) We use WeakRefs for callbacks and FinalizerRegistry to tell when
given callback is no longer referenced on the extension side. When we
detect that, we send message to UI side to remove it from the map.

---------

Co-authored-by: filip131311 <filip.kaminski@swmansiom.com>
Co-authored-by: Krzysztof Magiera <krzys@swmansion.com>