Move app root setting from vscode settings to launch configuration (#125)
This PR removes vscode setting that'd allow for specifying application
root using vscode setting.

We plan to use vscode's settings for visual / user customization rather
than project-specific parameters. The reason is that with vscode every
setting can be set globally and we can't limit them to be only
workspace-based. Given we already use launch configuration for
controlling build parameters, it makes sense to have things like appRoot
there.

As part of this I'm updating docs to include the new option, and also
updating troubleshooting guide where we mention the root can be change
in the vscode settings.