Add expo 51 support (#186)
This PR fixes issue with expo 51 and adds a test-app with expo go on SDK
51.

Here's what we're changing:
- we can no longer override runServer as it is now transpiled with
Object.defineProperty which cannot be modified. Instead we override
metro-config loadConfig method
- we used to report actual port number from runServer as well, since it
is no longer possible we patch node's http server module to report that
instead
- this change also affects non-expo apps, as we've been doing similar
overwrites for Metro.runServer which are no longer needed because of the
above change.