Add a way to close IDE, when in SidePanel. (#166)
Before: 
There was no way to close IDE, when in sidePanel, without switching to
tabPanel, or shutting device completely.

Solution: 
This PR adds a command "RNIDE.closePanel", that allows closing IDE Panel
in any location. Furthermore it adds a menu item to the side panel to
provide a way for triggering the command from UI.

Drawback: 
This solution removes IDE Side Panel From the UI completely until
reopening. Unfortunately vscode does not expose a way for disposing of
existing WebviewView.

Fixes: #146 

Realtes to:  #138


https://github.com/software-mansion/react-native-ide/assets/159789821/ba54d2b1-9dc6-4616-b0a6-d4db81f76d63

---------

Co-authored-by: filip131311 <filip.kaminski@swmansiom.com>