Remove import of expo-router from preview.js lib file (#211)
As spotted in #190 using preview functionality required you to have expo
router installed.

This was due to an import in lib file that gets loaded when you use the
preview package. The import was a legacy code that was put there when we
used a different way of loading preview package. Now that it is no
longer needed it can be safely removed.