Forward stderr lines to output channels used for iOS/Android builds (#251)
This PR fixes an issue where error messages printed to stderr from ios
or Android builds weren't included in the build output channels. The
reason was that we'd only listen for lines from stdout. This PR adds an
option for line reader to also include lines from stderr and we now use
this option for xcodebuild and gradle processes.