Use patched version of RN renderer for RN 0.74 to fix inspector
This PR implements a workaround for inspector issue on React Native 0.74
and Expo SDK 51.

On that versions, the inspector broke because of this commit
https://github.com/facebook/react/commit/37d901e2b8 in react core that
removed debug metadata from being passed down from createElement call
into react fiber node objects. Despite the problem being addressed by
using runtime based retireval approach in
https://github.com/facebook/react/commit/61bd00498 we couldn't really
use that as it doesn't provide the same level of detail about the
components. Specifically, metadata is missing for components w/o props
(which is frequent when used react-navigation or expo-router in React
Native), and the debug info would always point to component definition
line rather than the line where component is rendered. This turned out
to be a lot less practical.

There is a new approach implemented in
https://github.com/facebook/react/commit/151cce37401 but not yet
included in RN while also marked as experimental, so it isn't certain if
it has effects or can be used in later versions. Please checkout inline
comment in babel_transformer.js file for more context.

The fix this PR implements is to override renderer code that react
native pulls in. The renderer is build off of a certain version of
react-js and then commited into RN source code. Note that renderer code
is not kept in sync with react version used by react native. For
example, the mentioned commit that removes source metadata is not a part
of react release yet, while react-native depends on react 18.2. What we
ended up doing here, is that we pulled in the exact version of react,
the renderer is coming from in RN 74 (this is the version:
https://github.com/facebook/react/commit/03d6f7cf0), then build
renderers (separate files for new and old arch) with the mentioned
commit reverted. Then, we use babel_transformer to replace the renderer
by replacing source with the patched renderer versions that we are
adding to the repo.

This has been tested with RN 74 and Expo 51 projects.