Update sim-server to version that includes significant perf improvements (#362)
This PR pulls in new version of sim-server that includes
https://github.com/software-mansion-labs/simulator-server/pull/163

The linked PR significantly improves the performance of streaming
bringing CPU usage down for both the receiving end (panel) and the
server on its own. From initial testing it shows:
 - 40-50% less CPU usage on the renderer (IDE panel)
 - 80% less CPU usage on the server when streaming iOS simulator
 - 60% less CPU usage on the server when streaming Android emulator

On top of the above, this commit also includes some changes to the
sim-server build script that allows us to build modified version of
sim-server repo using build command. Before this change, when making
modifications to stream-server submodule, the script would check out the
version that's set in the parent repo. Now we only do that for Release
builds and in Debug, we show a prompt that allows us to just build the
modified version of the sim-server submodule.